requirements.txt
autogen-agentchat==0.6.4
autogen-ext[openai,azure]==0.6.4
python-dotenv
pydantic>=2.7.0
azure-cosmos>=4.5.0
aiohttp>=3.8.0
streamlit>=1.28.0